+++ wafer/app/viewer/widgets/combo_with_buttons.py
@@ -5,6 +5,11 @@ from ....core.qt.icon_engine import themed_icon
 from ....core.color.theme import ThemeManager
 
 
+class NoWheelComboBox(QComboBox):
+    def wheelEvent(self, event):
+        event.ignore()
+
+
 class ComboBoxWithButtons(QWidget):
     addClicked = Signal()
     removeClicked = Signal()
@@ -12,7 +17,7 @@ class ComboBoxWithButtons(QWidget):
 
     def __init__(self, parent=None):
         super().__init__(parent)
-        self.combo = QComboBox()
+        self.combo = NoWheelComboBox()
         self.button_add = QPushButton()
         self.button_add.setIcon(themed_icon("plus"))
         self.button_remove = QPushButton()